+++ lib/src/base/types/Dynamic.vue
@@ -56,6 +56,19 @@ export default {
               weight: 1
             }
       };
+
+      const queryStringObject = {}
+      if(Object.keys(this.$route.query).length !== 0) {
+        Object.keys(this.$route.query).map((queryKey) => {
+          if (queryKey.includes('[]')) {
+            queryStringObject[queryKey.substring(0, queryKey.indexOf('['))] = this.$route.query[queryKey].split(',')
+          } else queryStringObject[queryKey] = this.$route.query[queryKey]
+        })
+      }
+      if (this.$sections.queryStringSupport && this.$sections.queryStringSupport === "enabled") {
+        variables["query_string"] = queryStringObject
+      }
+
       const URL =
         `${this.$sections.serverUrl}/project/${this.$sections.projectId}/section/render`;
       this.$axios
@@ -80,11 +93,19 @@ export default {
           this.$emit("load", false);
         })
         .catch((e) => {
-          this.$emit('errorAddingSection', {
+          if (e.response.data.error) {
+            this.$emit('errorAddingSection', {
               closeModal: true,
               title: "Error adding "+ this.props.name,
-              message: "We couldn't save your changes, try again later"
+              message: e.response.data.error
             })
+          } else {
+            this.$emit('errorAddingSection', {
+              closeModal: true,
+              title: "Error adding "+ this.props.name,
+              message: this.$t('saveConfigSectionError')
+            })
+          }
 
           this.$emit("load", false);
         });